Split form stories into individual files
