+++ src/components/Form/Form.stories.mdx
@@ -1,5 +1,4 @@
 import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
-import * as stories from './Form.stories';
 
 <Meta title='Components/Form/Readme' />
 
@@ -444,9 +443,7 @@ const fields = useMemo(
 );
 ```
 <!-- ### Example
-<Preview withSource='none'>
-  <addressStories.Playground />
-</Preview> -->
+
 
 
 ## FormFieldAdvancedSelection
@@ -1697,10 +1694,6 @@ const fields = useMemo(
 );
 ```
 ## Form Example
-
-<Preview withSource='none'>
-  <stories.Playground />
-</Preview>
 
 ## Types
