+++ src/components/Form/stories/RuntimeBehaviours.stories.tsx
@@ -1,0 +1,119 @@
+import * as React from "react";
+import { ReactElement, useEffect, useMemo} from "react";
+import { withKnobs, boolean } from "@storybook/addon-knobs";
+
+// Utils
+import { useForm, formActions } from "@root/components/Form";
+import { validateEmail, validateSlow } from "../validators";
+import { renderButtons } from "@root/utils/storyUtils";
+
+// Components
+import Form from "../Form";
+
+// Types
+import { FieldDef } from "@root/components/Field";
+
+import { ORIGINAL_BODY_MARGIN } from "./utils";
+
+export default {
+	title: "Components/Form",
+	decorators: [withKnobs],
+};
+
+export const RuntimeBehaviours = (): ReactElement => {
+	const { state, dispatch } = useForm();
+
+	useEffect(() => {
+		document.body.style.margin = "0px";
+
+		return () => {
+			document.body.style.margin = ORIGINAL_BODY_MARGIN;
+		}
+	}, []);
+
+	const showState = boolean("Show state", false);
+
+	const fields = useMemo(
+		() : FieldDef[] =>
+			[
+				{
+					name: "text1",
+					label: "Simple Text",
+					type: "text",
+					instructionText: "testing",
+					validators: [validateEmail, validateSlow]
+				},
+				{
+					name: "text2",
+					label: "Text with validators and dynamic help",
+					type: "text",
+					helperText: state.data.text2,
+					validators: [validateEmail, validateSlow]
+				},
+				{
+					name: "text3",
+					label: "Text that copies to the next input",
+					type: "text"
+				},
+				{
+					name: "text4",
+					label: "Text that receives copy",
+					type: "text"
+				}
+			],
+		[]
+	);
+
+	useEffect(() => {
+		dispatch(
+			formActions.setFieldValue({
+				name: "text4",
+				value: state.data.text3
+			})
+		);
+	}, [state.data.text3]);
+
+	const setText1Value = function () {
+		dispatch(
+			formActions.setFieldValue({
+				name: "text1",
+				value: "test@test.com"
+			})
+		);
+	};
+
+	const setText2Value = function () {
+		dispatch(
+			formActions.setFieldValue({
+				name: "text2",
+				value: "notanemail"
+			})
+		);
+	};
+
+	return (
+		<>
+			{
+				showState && <pre>{JSON.stringify(state, null, "  ")}</pre>
+			}
+			<div style={{height: "100vh"}}>
+				<Form
+					buttons={renderButtons(dispatch)}
+					title='Runtime behaviors'
+					state={state}
+					fields={fields}
+					dispatch={dispatch}
+				/>
+			</div>
+			<div>
+				<p>
+					Here are some buttons that are not part of the form, but can change
+					values in the form proving communication between in/out of the form.
+					Notice that settext2 runs the validation after setting the value.
+				</p>
+				<button onClick={setText1Value}>Set Text1 Value</button>
+				<button onClick={setText2Value}>Set Text2 Value</button>
+			</div>
+		</>
+	);
+};
