+++ src/components/Form/stories/utils.tsx
@@ -1,0 +1,1 @@
+export const ORIGINAL_BODY_MARGIN = document.body.style.margin;
